+++ packages/web/src/Components/InputLabel.tsx
@@ -1,0 +1,4 @@
+import MaterialInputLabel from '@material-ui/core/InputLabel';
+export { InputLabelProps as IInputLabelProps } from '@material-ui/core/InputLabel';
+
+export default MaterialInputLabel;